fix: session keys
