+++ package.json
@@ -18,7 +18,7 @@
     "prepare": "husky"
   },
   "dependencies": {
-    "@argent/x-sessions": "7.0.0-beta.2",
+    "@argent/x-sessions": "7.0.0-beta.5",
     "@starknet-io/get-starknet-core": "4.0.4",
     "@starknet-react/chains": "^3.1.0",
     "@starknet-react/core": "^3.5.0",
@@ -26,7 +26,7 @@
     "react": "19.0.0-rc-02c0e824-20241028",
     "react-dom": "19.0.0-rc-02c0e824-20241028",
     "starknet": "^6.11.0",
-    "starknetkit": "^2.6.1"
+    "starknetkit": "2.6.2"
   },
   "devDependencies": {
     "@commitlint/cli": "^19.5.0",
