+++ src/constants/index.ts
@@ -6,8 +6,11 @@ export const ETHTokenAddress =
 export const DAITokenAddress =
   "0x00da114221cb83fa859dbdb4c44beeaa0bb37c7537ad5ae66fe5e0efd20e6eb3"
 
-export const ARGENT_DUMMY_CONTRACT_ADDRESS =
+export const ARGENT_DUMMY_CONTRACT_MAINNET_ADDRESS =
   "0x001c515f991f706039696a54f6f33730e9b0e8cc5d04187b13c2c714401acfd4"
+
+export const ARGENT_DUMMY_CONTRACT_SEPOLIA_ADDRESS =
+  "0x88d3cc4377a6cdfd27545a11548bd070c4e2e1e3df3d402922dbc4350b416"
 
 export const CHAIN_ID =
   process.env.NEXT_PUBLIC_CHAIN_ID === constants.NetworkName.SN_MAIN
@@ -36,3 +39,13 @@ export const ARGENT_SESSION_SERVICE_BASE_URL =
 export const ARGENT_WEBWALLET_URL =
   process.env.NEXT_PUBLIC_ARGENT_WEBWALLET_URL ||
   "https://sepolia-web.argent.xyz"
+
+export const USE_SEPOLIA_DUMMY_CONTRACT = process.env
+  .NEXT_PUBLIC_USE_SEPOLIA_DUMMY_CONTRACT
+  ? process.env.NEXT_PUBLIC_USE_SEPOLIA_DUMMY_CONTRACT === "true"
+  : false
+
+export const ARGENT_DUMMY_CONTRACT_ADDRESS =
+  CHAIN_ID === constants.NetworkName.SN_SEPOLIA
+    ? ARGENT_DUMMY_CONTRACT_SEPOLIA_ADDRESS
+    : ARGENT_DUMMY_CONTRACT_MAINNET_ADDRESS
